refactor: Move common implementation to runtime-common
