+++ runtime/common/Cargo.toml
@@ -3,15 +3,22 @@ name = "noir-runtime-common"
 version = "0.4.0"
 authors = ["Haderech Pte. Ltd."]
 edition = "2021"
-license = "Apache-2.0"
+license = "GPL-3.0-or-later"
 repository = "https://github.com/noirhq/noir.git"
 publish = false
 
 [dependencies]
+frame-support = { git = "https://github.com/paritytech/polkadot-sdk", branch = "stable2407", default-features = false }
+frame-system = { git = "https://github.com/paritytech/polkadot-sdk", branch = "stable2407", default-features = false }
 noir-core-primitives = { workspace = true }
+pallet-transaction-payment = { git = "https://github.com/paritytech/polkadot-sdk", branch = "stable2407", default-features = false }
+static_assertions = "1.1"
 
 [features]
 default = ["std"]
 std = [
+	"frame-support/std",
+	"frame-system/std",
 	"noir-core-primitives/std",
+	"pallet-transaction-payment/std",
 ]
